requirements.txt
streamlit==1.49.1
pandas==2.3.2
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
plotly==6.3.0
python-dotenv==1.1.1
